requirements.txt
asgiref==3.3.4
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
distro==1.5.0
Django==3.2.3
django-environ==0.4.5
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
ipython==7.25.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
packaging==20.9
paramiko==2.7.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
texttable==1.6.3
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
websocket-client==0.59.0
